requirements.txt
Django==3.2.6
django_autoslug==1.9.8
django_ckeditor==6.1.0
